Changed to use DAO pattern
