+++ java/src/META-INF/persistence.xml
@@ -12,7 +12,7 @@
 
         <!-- <class>EntityClassname</class>  -->
         <class>data.Topic</class>
-        <class>users.User</class>
+        <class>data.User</class>
 
         <properties>
             &login;
